Remove logs from toolbar
